requirements.txt
# FastAPI Backend
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Excel Processing
openpyxl==3.1.2
pandas==2.2.0

# AI/LLM
langchain==0.1.9
langchain-openai==0.0.5
langgraph==0.0.28
langchain-community==0.0.25

# Vector Store
chromadb==0.4.22

# Voice AI
openai==1.12.0
# elevenlabs==0.2.27  # Optional: for TTS

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Reporting
reportlab==4.0.9
markdown==3.5.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
